requirements.txt
azure-storage-blob>=12.14.0
azure-identity>=1.12.0
azure-mgmt-storage>=20.0.0
azure-mgmt-resource>=23.0.0
